requirements.txt
python-dotenv
docling
streamlit
firecrawl-py
langchain-community
langchain
langchain-pinecone
pinecone-notebooks
langchain-core
langchain-google-genai